Refactor - ThingDataManager를 사용하여 코드 리팩토링
